backend/requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-oidc==2.0.3
Flask-RESTful==0.3.10
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.20
numpy==1.26.2
pandas==2.2.0
debugpy==1.8.0
pyhumps==3.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1